// frontend/src/lib/api.d.ts
export interface components {
    schemas: {
// ...
        ConsumerCreate: {
            ack_policy?: components["schemas"]["AckPolicy"];
            /** @default 30 */
            ack_wait_seconds?: number;
            /** @default [] */
            backoff_seconds?: number[];
            deliver_group?: string | null;
            deliver_policy?: components["schemas"]["DeliverPolicy"];
            deliver_subject?: string | null;
            description?: string | null;
            /** @default true */
            durable?: boolean;
            /** @default [] */
            filter_subjects?: string[];
            /** @default false */
            flow_control?: boolean;
            /** @default false */
            headers_only?: boolean;
            /** @default 0 */
            idle_heartbeat_seconds?: number;
            /** @default 0 */
            inactive_threshold_seconds?: number;
            /** @default 1000 */
            max_ack_pending?: number;
            /** @default -1 */
            max_deliver?: number;
            /** @default 512 */
            max_waiting?: number;
            /** @default false */
            mem_storage?: boolean;
            metadata?: {
                [key: string]: string;
            };
            name: string;
            /** @default 0 */
            num_replicas?: number;
            opt_start_seq?: number | null;
            opt_start_time?: string | null;
            /** @default false */
            push?: boolean;
            /** @default 0 */
            rate_limit_bps?: number;
            replay_policy?: components["schemas"]["ReplayPolicy"];
            sample_freq?: string | null;
            stream: string;
        };
// ...
        ConsumerSummary: {
            ack_policy: components["schemas"]["AckPolicy"];
            ack_wait_seconds: number;
            /** @default [] */
            backoff_seconds?: number[];
            deliver_group?: string | null;
            deliver_policy: components["schemas"]["DeliverPolicy"];
            delivered_consumer_seq: number;
            delivered_stream_seq: number;
            description?: string | null;
            durable: boolean;
            filter_subjects: string[];
            health: components["schemas"]["ConsumerHealth"];
            lag: number | null;
            leader?: string | null;
            max_ack_pending: number;
            max_deliver: number;
            name: string;
            num_ack_pending: number;
            num_pending: number;
            num_redelivered: number;
            /** @default 0 */
            num_redelivered_total?: number;
            /** @default 0 */
            num_waiting?: number;
            /** @default false */
            paused?: boolean;
            paused_until?: string | null;
            push: boolean;
            stream: string;
        };
        /** ConsumerUpdate */
        ConsumerUpdate: {
            ack_wait_seconds?: number | null;
            backoff_seconds?: number[] | null;
            description?: string | null;
            filter_subjects?: string[] | null;
            headers_only?: boolean | null;
            inactive_threshold_seconds?: number | null;
            max_ack_pending?: number | null;
            max_deliver?: number | null;
            max_waiting?: number | null;
            metadata?: {
                [key: string]: string;
            } | null;
            rate_limit_bps?: number | null;
            sample_freq?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}